spi.v: make code synthetizable
Previously the code had negedge CS on sensitivity list, but then there
was a conditional that checked for high CS level. Negedge on the
sensitivity list was treated as reset signal, but the rest of the code
was reversed with respect to it. There are no logic elements capable of
implementing such logic, so the sensitivity list was changed to negedge
of logical sum of clock and CS. In addition, MISO was changed from
direct register output to wire gated by effective_cs, which as a side
effect reduced the number of warnings about tri-state logic produced by
Yosys.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>