WIP: spi.v: don't use wait states
Linux SPI TPM driver doesn't support it. This is an early attempt at
removing it. With changes to registers module that make accesses work
through combinational logic this "works" for registers other than
FIFO, most likely due to synchronous nature of RAM. Signals for
requesting read/write from registers are delayed through a chain of
LUTs, this is not portable. Length of that chain was obtained through
trials and errors. It may not work reliably on other units. Hardware
macros are not working in testbench started according to README.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>